Backend/requirements.txt
Flask
python-dotenv
pymongo
flask-cors
dnspython
